The easy fixes
