--- sbi/neural_nets/mdn_snpe_a.py
@@ -1,70 +1,0 @@
-# This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
-# under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
-
-from pyknos.mdn.mdn import MultivariateGaussianMDN
-from pyknos.nflows import transforms
-from torch import Tensor, nn
-
-from sbi.neural_nets.mog_flow_snpe_a import MoGFlow_SNPE_A
-from sbi.utils.sbiutils import standardizing_net, standardizing_transform
-
-
-def build_mdn_snpe_a(
-    batch_x: Tensor = None,
-    batch_y: Tensor = None,
-    z_score_x: bool = True,
-    z_score_y: bool = True,
-    hidden_features: int = 50,
-    num_components: int = 10,
-    embedding_net: nn.Module = nn.Identity(),
-    **kwargs,
-) -> nn.Module:
-    """Builds MDN p(x|y) with different sampling in training and evaluation mode.
-
-    Args:
-        batch_x: Batch of xs, used to infer dimensionality and (optional) z-scoring.
-        batch_y: Batch of ys, used to infer dimensionality and (optional) z-scoring.
-        z_score_x: Whether to z-score xs passing into the network.
-        z_score_y: Whether to z-score ys passing into the network.
-        hidden_features: Number of hidden features.
-        num_components: Number of components.
-        embedding_net: Optional embedding network for y.
-        kwargs: Additional arguments that are passed by the build function but are not
-            relevant for MDNs and are therefore ignored.
-
-    Returns:
-        Neural network.
-    """
-    x_numel = batch_x[0].numel()
-    # Infer the output dimensionality of the embedding_net by making a forward pass.
-    y_numel = embedding_net(batch_y[:1]).numel()
-
-    transform = transforms.IdentityTransform()
-
-    if z_score_x:
-        transform_zx = standardizing_transform(batch_x)
-        transform = transforms.CompositeTransform([transform_zx, transform])
-
-    if z_score_y:
-        embedding_net = nn.Sequential(standardizing_net(batch_y), embedding_net)
-
-    distribution = MultivariateGaussianMDN(
-        features=x_numel,
-        context_features=y_numel,
-        hidden_features=hidden_features,
-        hidden_net=nn.Sequential(
-            nn.Linear(y_numel, hidden_features),
-            nn.ReLU(),
-            nn.Dropout(p=0.0),
-            nn.Linear(hidden_features, hidden_features),
-            nn.ReLU(),
-            nn.Linear(hidden_features, hidden_features),
-            nn.ReLU(),
-        ),
-        num_components=num_components,
-        custom_initialization=True,
-    )
-
-    neural_net = MoGFlow_SNPE_A(transform, distribution, embedding_net)
-
-    return neural_net
